+++ lib/src/hardware/bosch.dart
@@ -1,0 +1,6 @@
+// Copyright (c) 2021, the Dart project authors.  Please see the AUTHORS file
+// for details. All rights reserved. Use of this source code is governed by a
+// BSD-style license that can be found in the LICENSE file.
+
+/// Oversampling multiplier defintion for BME280, BMP280 and BME680
+enum OversamplingMultiplier { X0, X1, X2, X4, X8, X16 }